// BE/shopdienmay/src/interfaces/product.interface.ts
export interface IProduct {
    name: string;
    description: string;
    image: string;
    imageList: string[];
    categories: [];
    price: number;
    sale_price: number;
}

// export interface IFile{
//     name: string;
//     image: string;
//     data: any;
//     size: number;
//     type: string;
// }